+++ crates/polars-io/Cargo.toml
@@ -141,7 +141,7 @@ cloud = [
   "reqwest",
   "http",
   "fastrand",
-  "chrono"
+  "chrono",
 ]
 file_cache = ["async", "dep:blake3", "dep:fs4", "dep:parking_lot", "serde_json", "cloud"]
 aws = ["object_store/aws", "cloud", "reqwest"]
