+++ crates/polars-io/src/cloud/mod.rs
@@ -20,6 +20,7 @@ pub use polars_object_store::*;
 pub mod cloud_writer;
 #[cfg(feature = "cloud")]
 pub mod concurrency;
+pub mod concurrency_config;
 #[cfg(feature = "cloud")]
 pub mod credential_provider;
