+++ crates/polars-plan/src/plans/conversion/dsl_to_ir/scans.rs
@@ -5,7 +5,7 @@ use either::Either;
 use polars_buffer::Buffer;
 use polars_core::runtime::ASYNC;
 use polars_io::RowIndex;
-use polars_io::cloud::FetchConfig;
+use polars_io::cloud::concurrency_config::FetchConfig;
 use polars_io::csv::read::streaming::read_until_start_and_infer_schema;
 use polars_io::prelude::*;
 use polars_io::utils::byte_source::{ByteSource, DynByteSourceBuilder};
