+++ crates/polars-stream/src/nodes/io_sources/multi_scan/functions/mod.rs
@@ -2,7 +2,8 @@ use std::num::NonZeroUsize;
 use std::sync::Arc;
 
 use polars_error::PolarsResult;
-use polars_io::cloud::{CloudOptions, FetchConfig};
+use polars_io::cloud::CloudOptions;
+use polars_io::cloud::concurrency_config::FetchConfig;
 use polars_io::metrics::IOMetrics;
 use polars_io::utils::byte_source::{ByteSource, DynByteSourceBuilder};
 use polars_io::utils::compression::SupportedCompression;
